fix(import): make separate component for import flow form.
