+++ frontend/public/extend/devconsole/components/overlay-import-flow/OverlayImportFlow.tsx
@@ -1,8 +1,6 @@
 import * as React from 'react';
 import { ModelessOverlay, Modal } from 'patternfly-react';
-import ImportFlow from '../import/import-flow';
-import { Firehose } from '../../../../components/utils';
-import { NamespaceModel } from '../../../../models';
+import { ImportFlow } from '../import/import-flow';
 import './OverlayImportFlow.scss';
 
 type Props = {
@@ -26,9 +24,7 @@ export const OverlayImportFlow: React.FunctionComponent<Props> = (props: Props) 
       </Modal.Header>
       <Modal.Body>
         <div className='odc-overlay-import__modal-body'>
-          <Firehose resources={[{kind: NamespaceModel.kind, prop: 'namespace', isList: true}]}>
-            <ImportFlow />
-          </Firehose>
+          <ImportFlow />
         </div>
       </Modal.Body>
     </ModelessOverlay>
